+++ src/test/java/model/PlayersTest.java
@@ -22,21 +22,4 @@ public class PlayersTest {
         assertThatThrownBy(() -> Players.from(names))
                 .isInstanceOf(IllegalArgumentException.class);
     }
-
-    @Test
-    @DisplayName("모든 플레이어에게 시작 카드를 분배한다.")
-    void 모든_플레이어에게_시작_카드들을_분배() {
-        //given
-        Deck deck = Deck.of();
-        Players players = new Players(List.of(
-                new Player("moda"),
-                new Player("daro"),
-                new Player("pobi")));
-        players.dealInitialCards(deck);
-
-        //when, then
-        for (Player player : players.getPlayers()) {
-            assertThat(player.getHandCards().size()).isEqualTo(2);
-        }
-    }
 }
